requirements.txt
# Core dependencies for SnapIndex
# Document processing
pypdf>=6.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-pptx>=0.6.21

# Audio processing and transcription
librosa>=0.10.0
sounddevice>=0.4.6
numpy>=1.24.0

# PyTorch and Whisper models
# torch>=2.0.0
openai-whisper>=20231117
scipy>=1.10.0

# ONNX runtime for Qualcomm models
onnxruntime>=1.16.0
onnx>=1.14.0
transformers>=4.35.0

# Qualcomm-specific packages (install if available)
# onnxruntime-qnn  # Uncomment if available via pip
# qnn-sdk  # Uncomment if available via pip

# Configuration and utilities
PyYAML>=6.0

# GUI framework
flet>=0.21.0

# Vector embeddings and FAISS
fastembed>=0.2.0
faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4  # Uncomment for GPU support

# Optional: For audio file loading
# audio2numpy>=0.1.5
# tiktoken>=0.5.0

# GUI (tkinter is included with Python, but listing for completeness)
# tkinter - built-in with Python

# Optional: For better ONNX performance on different platforms
# onnxruntime-openvino>=1.16.0  # For ARM/Intel optimization
# onnxruntime-directml>=1.16.0  # For Windows GPU acceleration

# Development and testing (optional)
# pytest>=7.0.0
